Review: give each decode step one owner
populateRawErrorMetadata was half-decoding the object itself (into a map it
only used for one key), decodeUpstreamError re-decoded the same bytes on the
fallback path, and the two were stitched together with a found-bool whose
value re-enumerated every struct field. The nested-vs-flat choice now lives
entirely in decodeUpstreamError, the per-field reads live in
upstreamErrorFromFields, and the bool protocol is gone: a zero-struct
comparison says whether the nested shape held anything.

Also gives http_error.go one canonical test file. The new tests land in
http_error_test.go instead of a bespoke http_error_status_test.go, and the
three parseHTTPErrorMetadata tests that had been living in responses_test.go
move in with them unchanged.

Mutation-checked again after the restructure: dropping the symbolic-status
routing or reinstating the all-or-nothing decode each fails the suite.
