+++ openai/responses_test.go
@@ -147,32 +147,6 @@ func TestResponsesAPI_ErrorMetadata(t *testing.T) {
 	assert.Equal(t, "prompt is too long", httpErr.Metadata.RawErrorMessage)
 }
 
-func TestParseHTTPErrorMetadata_FlatRawErrorWithNullNestedCode(t *testing.T) {
-	metadata := parseHTTPErrorMetadata(openAIErrorMetadata{
-		Raw: json.RawMessage(`{"message": "flat message", "type": "flat_type", "error": {"code": null}}`),
-	})
-
-	assert.Empty(t, metadata.RawErrorCode)
-	assert.Equal(t, "flat_type", metadata.RawErrorType)
-	assert.Equal(t, "flat message", metadata.RawErrorMessage)
-}
-
-func TestParseHTTPErrorMetadata_NestedStatusCodeOnly(t *testing.T) {
-	metadata := parseHTTPErrorMetadata(openAIErrorMetadata{
-		Raw: json.RawMessage(`{"error": {"status_code": 429}}`),
-	})
-
-	assert.Equal(t, 429, metadata.RawErrorStatusCode)
-}
-
-func TestParseHTTPErrorMetadata_NestedStatusOnly(t *testing.T) {
-	metadata := parseHTTPErrorMetadata(openAIErrorMetadata{
-		Raw: json.RawMessage(`{"error": {"status": 429}}`),
-	})
-
-	assert.Equal(t, 429, metadata.RawErrorStatusCode)
-}
-
 func TestResponsesStream_ReasoningOutputDoneSummaryFallback(t *testing.T) {
 	sse := strings.Join([]string{
 		`data: {"type":"response.created"}`,
